Rename catalog packages and refactor sync tests (#158)
* Make catalog package naming consistent with command flags

Rename:
* catalog/from-k8s -> catalog/to-consul
* catalog/from-consul -> catalog/to-k8s

* Refactor catalog/to-consul/resource_test.go

* Move common functionality into helper functions
* Remove duplicate tests
* Remove unnecessary time.Sleep calls

* Update catalog/to-consul/resource_test.go


* Add docs for test functions and change function names

* remove `test` prefix from helper function
* pass in `t *testing.T` object to helper function
  and assert NoError there rather than in the test
* make test IPs look more like IPs rather than string letters.
